mayor refactoring of checklist component
