+++ src/main/webapp/app/exam/manage/exam-management.module.ts
@@ -34,7 +34,8 @@ import { ProgrammingExerciseUtilsModule } from 'app/exercises/programming/shared
 import { TestRunManagementComponent } from 'app/exam/manage/test-runs/test-run-management.component';
 import { CreateTestRunModalComponent } from 'app/exam/manage/test-runs/create-test-run-modal.component';
 import { ArtemisMarkdownModule } from 'app/shared/markdown.module';
-import { ExamChecklistCheckComponent } from 'app/exam/manage/exams/exam-checklist-check/exam-checklist-check.component';
+import { ExamChecklistComponent } from 'app/exam/manage/exams/exam-checklist-component/exam-checklist.component';
+import { ExamChecklistCheckComponent } from 'app/exam/manage/exams/exam-checklist-component/exam-checklist-check/exam-checklist-check.component';
 
 const ENTITY_STATES = [...examManagementState];
 
@@ -77,6 +78,7 @@ const ENTITY_STATES = [...examManagementState];
         StudentExamDetailComponent,
         DurationPipe,
         StudentExamSummaryComponent,
+        ExamChecklistComponent,
         ExamChecklistCheckComponent,
     ],
 })
